move sql functions to service
