+++ cartoframes/data/observatory/enrichment/enrichment_service.py
@@ -179,6 +179,139 @@ class EnrichmentService(object):
 
         return project
 
+    def _get_points_enrichment_sql(self, temp_table_name, variables, filters):
+        tables_metadata = self._get_tables_metadata(variables).items()
+
+        return [self._build_points_query(table, metadata, temp_table_name, filters)
+                for table, metadata in tables_metadata]
+
+    def _build_points_query(self, table, metadata, temp_table_name, filters):
+        variables = ['enrichment_table.{}'.format(variable.column_name) for variable in metadata['variables']]
+        enrichment_dataset = metadata['dataset']
+        enrichment_geo_table = metadata['geo_table']
+        data_table = '{project}.{user_dataset}.{temp_table_name}'.format(
+            project=self.working_project,
+            user_dataset=self.user_dataset,
+            temp_table_name=temp_table_name
+        )
+
+        return '''
+            SELECT data_table.{enrichment_id}, {variables},
+                ST_Area(enrichment_geo_table.geom) AS {table}_area
+            FROM `{enrichment_dataset}` enrichment_table
+                JOIN `{enrichment_geo_table}` enrichment_geo_table
+                    ON enrichment_table.geoid = enrichment_geo_table.geoid
+                JOIN `{data_table}` data_table
+                    ON ST_Within(data_table.{geojson_column}, enrichment_geo_table.geom)
+            {where};
+        '''.format(
+            variables=', '.join(variables),
+            geojson_column=self.geojson_column,
+            enrichment_dataset=enrichment_dataset,
+            enrichment_geo_table=enrichment_geo_table,
+            enrichment_id=self.enrichment_id,
+            where=self._build_where_clausule(filters),
+            data_table=data_table,
+            table=table
+        )
+
+    def _get_polygon_enrichment_sql(self, temp_table_name, variables, filters, aggregation):
+        tables_metadata = self._get_tables_metadata(variables).items()
+
+        return [self._build_polygons_query(table, metadata, temp_table_name, filters, aggregation)
+                for table, metadata in tables_metadata]
+
+    def _build_polygons_query(self, table, metadata, temp_table_name, filters, aggregation):
+        variables = metadata['variables']
+        enrichment_dataset = metadata['dataset']
+        enrichment_geo_table = metadata['geo_table']
+        data_table = '{project}.{user_dataset}.{temp_table_name}'.format(
+            project=self.working_project,
+            user_dataset=self.user_dataset,
+            temp_table_name=temp_table_name
+        )
+
+        if aggregation == AGGREGATION_NONE:
+            grouper = ''
+            variables = self._build_polygons_query_variables_without_aggregation(variables)
+        else:
+            grouper = 'group by data_table.{enrichment_id}'.format(enrichment_id=self.enrichment_id)
+            variables = self._build_polygons_query_variables_with_aggregation(variables, aggregation)
+
+        return '''
+            SELECT data_table.{enrichment_id}, {variables}
+            FROM `{enrichment_dataset}` enrichment_table
+                JOIN `{enrichment_geo_table}` enrichment_geo_table
+                    ON enrichment_table.geoid = enrichment_geo_table.geoid
+                JOIN `{data_table}` data_table
+                    ON ST_Intersects(data_table.{geojson_column}, enrichment_geo_table.geom)
+            {where}
+            {grouper};
+        '''.format(
+                geojson_column=self.geojson_column,
+                enrichment_dataset=enrichment_dataset,
+                enrichment_geo_table=enrichment_geo_table,
+                enrichment_id=self.enrichment_id,
+                where=self._build_where_clausule(filters),
+                data_table=data_table,
+                grouper=grouper or '',
+                variables=variables
+            )
+
+    def _build_polygons_query_variables_with_aggregation(self, variables, aggregation):
+        return ', '.join([
+            self._build_polygons_query_variable_with_aggregation(
+                variable,
+                aggregation
+            ) for variable in variables])
+
+    def _build_polygons_query_variable_with_aggregation(self, variable, aggregation):
+        variable_agg = _get_aggregation(variable, aggregation)
+
+        if (variable_agg == 'SUM'):
+            return """
+                {aggregation}(
+                    enrichment_table.{column} * (
+                        ST_Area(ST_Intersection(enrichment_geo_table.geom, data_table.{geo_column}))
+                        /
+                        ST_area(data_table.{geo_column})
+                    )
+                ) AS {aggregation}_{column}
+                """.format(
+                    column=variable.column_name,
+                    geo_column=self.geojson_column,
+                    aggregation=variable_agg)
+        else:
+            return """
+                {aggregation}(
+                    enrichment_table.{column} * (
+                        ST_Area(ST_Intersection(enrichment_geo_table.geom, data_table.{geo_column}))
+                    )
+                ) AS {aggregation}_{column}
+                """.format(
+                    column=variable.column_name,
+                    geo_column=self.geojson_column,
+                    aggregation=variable_agg)
+
+    def _build_polygons_query_variables_without_aggregation(self, variables):
+        variables = ['enrichment_table.{}'.format(variable.column_name) for variable in variables]
+
+        return """
+            {variables},
+            ST_Area(ST_Intersection(enrichment_geo_table.geom, data_table.{geojson_column})) /
+            ST_area(data_table.{geojson_column}) AS measures_proportion
+            """.format(
+                variables=', '.join(variables),
+                geojson_column=self.geojson_column)
+
+    def _build_where_clausule(self, filters):
+        where = ''
+        if len(filters) > 0:
+            where_clausules = ["enrichment_table.{} {}".format(f.variable.column_name, f.query) for f in filters]
+            where = 'WHERE {}'.format('AND '.join(where_clausules))
+
+        return where
+
 
 def prepare_variables(variables, aggregation=None):
     if isinstance(variables, list):
